requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
httpx==0.27.0
beautifulsoup4
lxml
python-dotenv
pydantic>=2.7
duckduckgo_search==5.3.1
Pillow
torch==2.2.2
torchvision==0.17.2
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git